rename frontend dir to jfr; fix flame-graph component bug: if no data, siblings[] is not cleared
