+++ frontend/src/components/jfr/Toolbar.vue
