+++ frontend/src/components/jfr/utils.ts
