+++ frontend/src/i18n/jfr/en.ts
