+++ frontend/src/i18n/jfr/zh.ts
